refactor(vitest): Split Vitest matcher tests
