+++ packages/vitest-file-snapshots/custom-data/validation/src/matcher-configuration/stores_snapshots_in_custom_directories.txt
