+++ packages/vitest-file-snapshots/data/test/validation/matcher-configuration/resolves_test_path_relative_to_testDir.txt
