+++ packages/vitest-file-snapshots/data/test/validation/src/json-file-matcher/applies-custom-file-path-resolver/name.json
