+++ packages/vitest-file-snapshots/data/test/validation/src/json-file-matcher/applies_name/value_1.json
