+++ packages/vitest-file-snapshots/data/test/validation/src/json-file-matcher/when_includeUndefinedObjectProperties_is_true_serializes_undefined_object_properties.json
