+++ packages/vitest-file-snapshots/data/test/validation/src/matcher-configuration/applies-custom-file-path-resolver/name.txt
