+++ packages/vitest-file-snapshots/data/test/validation/src/matcher-configuration/applies_indentSize_to_JSON_file_snapshots.json
