+++ packages/vitest-file-snapshots/data/test/validation/src/matcher-configuration/ignores_indentSize_in_text_file_snapshots.txt
