+++ packages/vitest-file-snapshots/data/test/validation/src/text-file-matcher/applies-custom-file-path-resolver/name.txt
