+++ packages/vitest-file-snapshots/data/test/validation/src/text-file-matcher/applies_custom_file_extension.md
