+++ packages/vitest-file-snapshots/data/test/validation/src/text-file-matcher/applies_name/value_1.txt
