+++ packages/vitest-file-snapshots/data/test/validation/src/text-file-matcher/matches_value_with_text_file.txt
